Fixed some imports and split yaml_helper into an executable and a moudle
