+++ scripts/data_processing_pipeline.py
@@ -13,8 +13,8 @@ import shutil
 from job_submit import submit_slurm
 from mwa_metadb_utils import get_common_obs_metadata
 import submit_to_database as std
-import yaml_helper
-import dpp_check_args
+from dpp import yaml_helper
+from dpp import dpp_check_args
 
 
 comp_config = load_config_file()
