+++ setup.py
@@ -60,8 +60,8 @@ setup(name="mwa_search",
                'scripts/splice_wrapper.py', 'scripts/LOTAAS_wrapper.py',
                'scripts/search_launch_loop.sh', 'scripts/rsync_rm_loop.sh',
                'scripts/bestgridpos.py', 'scripts/data_processing_pipeline.py',
-               'dpp_scripts/stokes_fold.py', 'dpp_scripts/yaml_helper.py',
-               'dpp_scripts/pulsars_in_fov.py',
+               'dpp_scripts/stokes_fold.py', 'dpp_scripts/make_pulsar_yaml.py',
+               'dpp_scripts/pulsars_in_fov.py', 'dpp_scripts/prepfold_cmd_make.py',
                'plotting/plot_obs_pulsar.py',
                'plotting/position_sn_heatmap_fwhm.py',
                'dpp/plotting_toolkit.py',
